backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.8
alembic==1.12.1
pydantic==2.4.2
pandas==2.1.3
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
groq==0.4.1
sentence-transformers==2.2.2
faiss-cpu==1.7.4
numpy==1.24.3
python-multipart==0.0.6
python-dotenv==1.0.0
httpx==0.25.2
aiofiles==23.2.1